+++ prog/dftb+/lib_dftbplus/main.F90
@@ -64,7 +64,11 @@ module dftbp_main
   use dftbp_elecconstraints
   use dftbp_pmlocalisation, only : TPipekMezey
   use dftbp_linresp
+  use dftbp_linresptypes
   use dftbp_pprpa, only : ppRPAenergies
+#:if WITH_ARPACK
+  use dftbp_RS_LinearResponse
+#:endif
   use dftbp_mainio
   use dftbp_commontypes
   use dftbp_dispersions, only : TDispersionIface
@@ -167,7 +171,7 @@ contains
 
     ! Main geometry loop
     geoOpt: do iGeoStep = 0, nGeoSteps
-      tWriteRestart = env%tGlobalMaster&
+      tWriteRestart = env%tGlobalLead&
           & .and. needsRestartWriting(isGeoOpt, tMd, iGeoStep, nGeoSteps, restartFreq)
       call printGeoStepInfo(tCoordOpt, tLatOpt, iLatGeoStep, iGeoStep)
       call processGeometry(env, iGeoStep, iLatGeoStep, tWriteRestart, tStopDriver, tStopScc,&
@@ -217,7 +221,7 @@ contains
     call env%globalTimer%startTimer(globalTimers%postGeoOpt)
 
   #:if WITH_SOCKETS
-    if (tSocket .and. env%tGlobalMaster) then
+    if (tSocket .and. env%tGlobalLead) then
       call socket%shutdown()
     end if
   #:endif
@@ -228,7 +232,7 @@ contains
 
     tGeomEnd = tMD .or. tGeomEnd .or. tDerivs
 
-    if (env%tGlobalMaster) then
+    if (env%tGlobalLead) then
       if (tWriteDetailedOut) then
         call writeDetailedOut5(fdDetailedOut, isGeoOpt, tGeomEnd, tMd, tDerivs, tEField, absEField,&
             & dipoleMoment)
@@ -241,7 +245,7 @@ contains
       end if
     end if
 
-    if (env%tGlobalMaster .and. tDerivs) then
+    if (env%tGlobalLead .and. tDerivs) then
       call getHessianMatrix(derivDriver, pDynMatrix)
       call writeHessianOut(hessianOut, pDynMatrix)
     else
@@ -310,9 +314,9 @@ contains
           & lCurrArray)
     end if
     if (tWriteResultsTag) then
-      call writeResultsTag(resultsTag, energy, derivs, chrgForces, electronicSolver, tStress,&
-          & totalStress, pDynMatrix, tPeriodic, cellVol, tMulliken, qOutput, q0, taggedWriter,&
-          & tDefinedFreeE, cm5Cont)
+      call writeResultsTag(resultsTag, energy, derivs, chrgForces, nEl, Ef, eigen, filling,&
+          & electronicSolver, tStress, totalStress, pDynMatrix, tPeriodic, cellVol, tMulliken,&
+          & qOutput, q0, taggedWriter, tDefinedFreeE, cm5Cont)
     end if
     if (tWriteDetailedXML) then
       call writeDetailedXml(runId, speciesName, species0, pCoord0Out, tPeriodic, tHelical, latVec,&
@@ -529,7 +533,7 @@ contains
 
         ! check charge convergece and guess new eigenvectors
         tStopScc = hasStopFile(fStopScc)
-        if (tRangeSep) then
+        if (isRangeSep) then
           call getReksNextInputDensity(sccErrorQ, sccTol, tConverged, iSccIter, minSccIter,&
               & maxSccIter, iGeoStep, tStopScc, eigvecsReal, deltaRhoOut, deltaRhoIn, deltaRhoDiff,&
               & reks)
@@ -562,7 +566,7 @@ contains
                 & pCoord0Out, q0, qOutput, orb, species, tPrintMulliken, extPressure, cellVol,&
                 & tAtomicEnergy, tDispersion, tPeriodic, tSccCalc, invLatVec, kPoint,&
                 & iAtInCentralRegion, electronicSolver, tDefinedFreeE, reks, allocated(thirdOrd),&
-                & tRangeSep)
+                & isRangeSep)
           end if
           if (tWriteBandDat) then
             call writeBandOut(bandOut, eigen, filling, kWeight)
@@ -640,7 +644,7 @@ contains
             & deltaRhoInSqr, deltaRhoOutSqr, qOutput, nNeighbourLC, tLargeDenseMatrices)
 
         !> For rangeseparated calculations deduct atomic charges from deltaRho
-        if (tRangeSep) then
+        if (isRangeSep) then
           select case(nSpin)
           case(2)
             do iSpin = 1, 2
@@ -711,7 +715,7 @@ contains
 
         ! Mix charges Input/Output
         if (tSccCalc) then
-          if(.not. tRangeSep) then
+          if(.not. isRangeSep) then
             call getNextInputCharges(env, pChrgMixer, qOutput, qOutRed, orb, nIneqOrb, iEqOrbitals,&
                 & iGeoStep, iSccIter, minSccIter, maxSccIter, sccTol, tStopScc, tMixBlockCharges,&
                 & tReadChrg, qInput, qInpRed, sccErrorQ, tConverged, qBlockOut, iEqBlockDftbU,&
@@ -735,7 +739,7 @@ contains
             call printBlankLine()
           end if
 
-          tWriteSccRestart = env%tGlobalMaster .and. &
+          tWriteSccRestart = env%tGlobalLead .and. &
               & needsSccRestartWriting(restartFreq, iGeoStep, iSccIter, minSccIter, maxSccIter,&
               & tMd, isGeoOpt, tDerivs, tConverged, tReadChrg, tStopScc)
           if (tWriteSccRestart) then
@@ -753,7 +757,7 @@ contains
               & E0, extPressure, cellVol, tAtomicEnergy, tDispersion, tEField, tPeriodic, nSpin,&
               & tSpin, tSpinOrbit, tSccCalc, allocated(onSiteElements), tNegf, invLatVec, kPoint,&
               & iAtInCentralRegion, electronicSolver, tDefinedFreeE, allocated(halogenXCorrection),&
-              & tRangeSep, allocated(thirdOrd), allocated(solvation), cm5Cont)
+              & isRangeSep, allocated(thirdOrd), allocated(solvation), cm5Cont)
         end if
 
         if (tConverged .or. tStopScc) then
@@ -773,19 +777,21 @@ contains
     call env%globalTimer%startTimer(globalTimers%postSCC)
 
     if (isLinResp) then
-      if (withMpi) then
-        call error("Linear response calc. does not work with MPI yet")
+      if (.not. isRS_LinResp) then
+        call calculateLinRespExcitations(env, linearResponse, parallelKS, sccCalc, qOutput, q0,&
+            & over, eigvecsReal, eigen(:,1,:), filling(:,1,:), coord, species, speciesName, orb,&
+            & skHamCont, skOverCont, autotestTag, taggedWriter, runId, neighbourList,&
+            & nNeighbourSK, denseDesc, iSparseStart, img2CentCell, tWriteAutotest, tCasidaForces,&
+            & tLinRespZVect, tPrintExcitedEigvecs, tPrintEigvecsTxt, nonSccDeriv, energy,&
+            & energiesCasida, SSqrReal, rhoSqrReal, excitedDerivs, dQAtomEx, occNatural)
+      else
+        call calculateLinRespExcitations_RS(env, linearResponse, parallelKS, sccCalc, qOutput, q0,&
+            & over, eigvecsReal, eigen(:,1,:), filling(:,1,:), coord0, species, speciesName, orb,&
+            & skHamCont, skOverCont, autotestTag, taggedWriter, runId, neighbourList, nNeighbourSK,&
+            & denseDesc, iSparseStart, img2CentCell, tWriteAutotest, tCasidaForces, tLinRespZVect,&
+            & tPrintExcitedEigvecs, tPrintEigvecsTxt, nonSccDeriv, energy, energiesCasida,&
+            & SSqrReal, deltaRhoOutSqr, excitedDerivs, dQAtomEx, occNatural, rangeSep)
       end if
-      if (allocated(solvation)) then
-        call error("Solvation model do not work with linear response yet.")
-      end if
-      call ensureLinRespConditions(t3rd, tRealHS, tPeriodic, tCasidaForces)
-      call calculateLinRespExcitations(env, lresp, parallelKS, sccCalc, qOutput, q0, over,&
-          & eigvecsReal, eigen(:,1,:), filling(:,1,:), coord, species, speciesName, orb, tHelical,&
-          & skHamCont, skOverCont, autotestTag, taggedWriter, runId, neighbourList, nNeighbourSK,&
-          & denseDesc, iSparseStart, img2CentCell, tWriteAutotest, tCasidaForces, tLinRespZVect,&
-          & tPrintExcitedEigvecs, tPrintEigvecsTxt, nonSccDeriv, energy, energiesCasida, SSqrReal,&
-          & rhoSqrReal, excitedDerivs, occNatural)
     end if
 
     if (allocated(ppRPA)) then
@@ -808,10 +814,10 @@ contains
 
     if (tDipole .and. .not.allocated(reks)) then
       call getDipoleMoment(qOutput, q0, coord, dipoleMoment, iAtInCentralRegion)
-    #:call DEBUG_CODE
+    #:block DEBUG_CODE
       call checkDipoleViaHellmannFeynman(rhoPrim, q0, coord0, over, orb, neighbourList,&
           & nNeighbourSk, species, iSparseStart, img2CentCell)
-    #:endcall DEBUG_CODE
+    #:endblock DEBUG_CODE
     end if
 
     call env%globalTimer%startTimer(globalTimers%eigvecWriting)
@@ -1506,7 +1512,7 @@ contains
 
     integer :: nSpin
 
-    #:block ASSERT_CODE
+    #:block DEBUG_CODE
       @:ASSERT(size(H0) == size(over))
       if (.not. allocated(reks)) then
         @:ASSERT(size(ham, dim=1) == size(over))
@@ -1519,7 +1525,7 @@ contains
           @:ASSERT(size(ERhoPrim) == size(rhoPrim, dim=1))
         end if
       end if
-    #:endblock ASSERT_CODE
+    #:endblock DEBUG_CODE
 
     if (allocated(reks)) then
       if (size(over, dim=1) == sparseSize) then
@@ -1934,12 +1940,12 @@ contains
 
   !> Returns the density matrix using dense diagonalisation.
   subroutine getDensityFromDenseDiag(env, denseDesc, ham, over, neighbourList, nNeighbourSK,&
-      & iSparseStart, img2CentCell, iCellVec, cellVec, kPoint, kWeight, orb, iAtomStart, tHelical, coord,&
-      & species, electronicSolver, tRealHS, tSpinSharedEf, tSpinOrbit, tDualSpinOrbit, tFillKSep,&
-      & tFixEf, tMulliken, iDistribFn, tempElec, nEl, parallelKS, Ef, energy, rangeSep, eigen,&
-      & filling, rhoPrim, Eband, TS, E0, iHam, xi, orbitalL, HSqrReal, SSqrReal, eigvecsReal,&
-      & iRhoPrim, HSqrCplx, SSqrCplx, eigvecsCplx, rhoSqrReal, deltaRhoInSqr, deltaRhoOutSqr,&
-      & qOutput, nNeighbourLC)
+      & iSparseStart, img2CentCell, iCellVec, cellVec, kPoint, kWeight, orb, iAtomStart, tHelical,&
+      & coord, species, electronicSolver, tRealHS, tSpinSharedEf, tSpinOrbit, tDualSpinOrbit,&
+      & tFillKSep, tFixEf, tMulliken, iDistribFn, tempElec, nEl, parallelKS, Ef, energy, rangeSep,&
+      & eigen, filling, rhoPrim, Eband, TS, E0, iHam, xi, orbitalL, HSqrReal, SSqrReal,&
+      & eigvecsReal, iRhoPrim, HSqrCplx, SSqrCplx, eigvecsCplx, rhoSqrReal, deltaRhoInSqr,&
+      & deltaRhoOutSqr, qOutput, nNeighbourLC)
 
     !> Environment settings
     type(TEnvironment), intent(inout) :: env
@@ -2270,7 +2276,7 @@ contains
 
       ! Add rangeseparated contribution
       ! Assumes deltaRhoInSqr only used by rangeseparation
-      ! Should this be used elsewhere, need to pass tRangeSep
+      ! Should this be used elsewhere, need to pass isRangeSep
       if (allocated(rangeSep)) then
         call denseMulliken(deltaRhoInSqr, SSqrReal, denseDesc%iAtomStart, qOutput)
         call rangeSep%addLRHamiltonian(env, deltaRhoInSqr(:,:,iSpin), over,&
@@ -3669,48 +3675,19 @@ contains
   end function needsSccRestartWriting
 
 
-  !> Stop if linear response module can not be invoked due to unimplemented combinations of
-  !> features.
-  subroutine ensureLinRespConditions(t3rd, tRealHS, tPeriodic, tForces)
-
-    !> 3rd order hamiltonian contributions included
-    logical, intent(in) :: t3rd
-
-    !> a real hamiltonian
-    logical, intent(in) :: tRealHs
-
-    !> periodic boundary conditions
-    logical, intent(in) :: tPeriodic
-
-    !> forces being evaluated in the excited state
-    logical, intent(in) :: tForces
-
-    if (t3rd) then
-      call error("Third order currently incompatible with excited state")
-    end if
-    if (.not. tRealHS) then
-      call error("Only real systems are supported for excited state calculations")
-    end if
-    if (tPeriodic .and. tForces) then
-      call error("Forces in the excited state for periodic geometries are currently unavailable")
-    end if
-
-  end subroutine ensureLinRespConditions
-
-
   !> Do the linear response excitation calculation.
-  subroutine calculateLinRespExcitations(env, lresp, parallelKS, sccCalc, qOutput, q0, over,&
-      & eigvecsReal, eigen, filling, coord, species, speciesName, orb, tHelical, skHamCont,&
+  subroutine calculateLinRespExcitations(env, linearResponse, parallelKS, sccCalc, qOutput, q0,&
+      & over, eigvecsReal, eigen, filling, coord, species, speciesName, orb, skHamCont,&
       & skOverCont, autotestTag, taggedWriter, runId, neighbourList, nNeighbourSk, denseDesc,&
       & iSparseStart, img2CentCell, tWriteAutotest, tForces, tLinRespZVect, tPrintExcEigvecs,&
       & tPrintExcEigvecsTxt, nonSccDeriv, energy, energies, work, rhoSqrReal, excitedDerivs,&
-      & occNatural)
+      & dQAtomEx, occNatural)
 
     !> Environment settings
     type(TEnvironment), intent(in) :: env
 
     !> excited state settings
-    type(TLinResp), intent(inout) :: lresp
+    type(TLinResp), intent(inout), allocatable :: linearResponse
 
     !> K-points and spins to process
     type(TParallelKS), intent(in) :: parallelKS
@@ -3747,9 +3724,6 @@ contains
 
     !> Atomic orbital information
     type(TOrbitals), intent(in) :: orb
-
-    !> Is the geometry helical
-    logical, intent(in) :: tHelical
 
     !> non-SCC hamiltonian information
     type(TSlakoCont), intent(in) :: skHamCont
@@ -3814,6 +3788,9 @@ contains
     !> excited state energy derivative with respect to atomic coordinates
     real(dp), intent(inout), allocatable :: excitedDerivs(:,:)
 
+    !> Mulliken charges in excited state
+    real(dp), intent(out) :: dQAtomEx(:)
+
     !> natural orbital occupation numbers
     real(dp), intent(inout), allocatable :: occNatural(:)
 
@@ -3831,13 +3808,8 @@ contains
     energy%Eexcited = 0.0_dp
     allocate(dQAtom(nAtom))
     dQAtom(:) = sum(qOutput(:,:,1) - q0(:,:,1), dim=1)
-    if (tHelical) then
-      call unpackHelicalHS(work, over, neighbourList%iNeighbour, nNeighbourSK,&
-          & denseDesc%iAtomStart, iSparseStart, img2CentCell, orb, species, coord)
-    else
-      call unpackHS(work, over, neighbourList%iNeighbour, nNeighbourSK, denseDesc%iAtomStart,&
-          & iSparseStart, img2CentCell)
-    end if
+    call unpackHS(work, over, neighbourList%iNeighbour, nNeighbourSK, denseDesc%iAtomStart,&
+        & iSparseStart, img2CentCell)
     call blockSymmetrizeHS(work, denseDesc%iAtomStart)
     if (allocated(rhoSqrReal)) then
       do iSpin = 1, nSpin
@@ -3852,19 +3824,20 @@ contains
       if (tPrintExcEigVecs) then
         allocate(naturalOrbs(orb%nOrb, orb%nOrb, 1))
       end if
-      call addGradients(tSpin, lresp, denseDesc%iAtomStart, eigvecsReal, eigen, work, filling,&
-          & coord(:,:nAtom), sccCalc, dQAtom, pSpecies0, neighbourList%iNeighbour, img2CentCell,&
-          & orb, skHamCont, skOverCont, tWriteAutotest, fdAutotest, taggedWriter, energy%Eexcited,&
-         & energies, excitedDerivs, nonSccDeriv, rhoSqrReal, occNatural, naturalOrbs)
+      call addGradients(tSpin, linearResponse, denseDesc%iAtomStart, eigvecsReal, eigen, work,&
+          & filling, coord(:,:nAtom), sccCalc, dQAtom, pSpecies0, neighbourList%iNeighbour,&
+          & img2CentCell, orb, skHamCont, skOverCont, tWriteAutotest, fdAutotest, taggedWriter,&
+          & energy%Eexcited, energies, excitedDerivs, nonSccDeriv, rhoSqrReal, occNatural,&
+          & naturalOrbs)
       if (tPrintExcEigvecs) then
         call writeRealEigvecs(env, runId, neighbourList, nNeighbourSK, denseDesc, iSparseStart,&
             & img2CentCell, pSpecies0, speciesName, orb, over, parallelKS, tPrintExcEigvecsTxt,&
             & naturalOrbs, work, fileName="excitedOrbs")
       end if
     else
-      call calcExcitations(lresp, tSpin, denseDesc, eigvecsReal, eigen, work, filling,&
-          & coord(:,:nAtom), sccCalc, dQAtom, pSpecies0, neighbourList%iNeighbour, img2CentCell,&
-          & orb, tWriteAutotest, fdAutotest, taggedWriter, energy%Eexcited, energies, rhoSqrReal)
+      call linResp_calcExcitations(linearResponse, tSpin, denseDesc, eigvecsReal, eigen, work,&
+          & filling, coord(:,:nAtom), sccCalc, dQAtom, pSpecies0, neighbourList%iNeighbour,&
+          & img2CentCell, orb, tWriteAutotest, fdAutotest, taggedWriter, energy%Eexcited, energies)
     end if
     energy%Etotal = energy%Etotal + energy%Eexcited
     energy%EMermin = energy%EMermin + energy%Eexcited
@@ -3875,6 +3848,207 @@ contains
 
   end subroutine calculateLinRespExcitations
 
+
+  !> Do the linear response excitation calculation with range-separated Hamiltonian.
+  subroutine calculateLinRespExcitations_RS(env, linearResponse, parallelKS, sccCalc, qOutput, q0,&
+      & over, eigvecsReal, eigen, filling, coord0, species, speciesName, orb, skHamCont,&
+      & skOverCont, autotestTag, taggedWriter, runId, neighbourList, nNeighbourSk, denseDesc,&
+      & iSparseStart, img2CentCell, tWriteAutotest, tForces, tLinRespZVect, tPrintExcEigvecs,&
+      & tPrintExcEigvecsTxt, nonSccDeriv, energy, energies, work, deltaRhoOutSqr, excitedDerivs,&
+      & dQAtomEx, occNatural, rangeSep)
+
+    !> Environment settings
+    type(TEnvironment), intent(in) :: env
+
+    !> excited state settings
+    type(TLinResp), intent(inout) :: linearResponse
+
+    !> K-points and spins to process
+    type(TParallelKS), intent(in) :: parallelKS
+
+    !> SCC module internal variables
+    type(TScc), intent(in) :: sccCalc
+
+    !> electrons in atomic orbitals
+    real(dp), intent(in) :: qOutput(:,:,:)
+
+    !> reference atomic orbital occupations
+    real(dp), intent(in) :: q0(:,:,:)
+
+    !> sparse overlap matrix
+    real(dp), intent(in) :: over(:)
+
+    !> ground state eigenvectors
+    real(dp), intent(inout) :: eigvecsReal(:,:,:)
+
+    !> ground state eigenvalues (orbital, kpoint)
+    real(dp), intent(in) :: eigen(:,:)
+
+    !> ground state fillings (orbital, kpoint)
+    real(dp), intent(in) :: filling(:,:)
+
+    !> central cell coordinates
+    real(dp), intent(in) :: coord0(:,:)
+
+    !> species of all atoms in the system
+    integer, target, intent(in) :: species(:)
+
+    !> label for each atomic chemical species
+    character(*), intent(in) :: speciesName(:)
+
+    !> Atomic orbital information
+    type(TOrbitals), intent(in) :: orb
+
+    !> non-SCC hamiltonian information
+    type(TSlakoCont), intent(in) :: skHamCont
+
+    !> overlap information
+    type(TSlakoCont), intent(in) :: skOverCont
+
+    !> File name for regression data
+    character(*), intent(in) :: autotestTag
+
+    !> Tagged writer
+    type(TTaggedWriter), intent(inout) :: taggedWriter
+
+    !> Job ID for future identification
+    integer, intent(in) :: runId
+
+    !> list of neighbours for each atom
+    type(TNeighbourList), intent(in) :: neighbourList
+
+    !> Number of neighbours for each of the atoms
+    integer, intent(in) :: nNeighbourSK(:)
+
+    !> Dense matrix descriptor
+    type(TDenseDescr), intent(in) :: denseDesc
+
+    !> Index array for the start of atomic blocks in sparse arrays
+    integer, intent(in) :: iSparseStart(:,:)
+
+    !> map from image atoms to the original unique atom
+    integer, intent(in) :: img2CentCell(:)
+
+    !> should regression test data be written
+    logical, intent(in) :: tWriteAutotest
+
+    !> forces to be calculated in the excited state
+    logical, intent(in) :: tForces
+
+    !> require the Z vector for excited state properties
+    logical, intent(in) :: tLinRespZVect
+
+    !> print natural orbitals of the excited state
+    logical, intent(in) :: tPrintExcEigvecs
+
+    !> print natural orbitals also in text form?
+    logical, intent(in) :: tPrintExcEigvecsTxt
+
+    !> method for calculating derivatives of S and H0
+    type(TNonSccDiff), intent(in) :: nonSccDeriv
+
+    !> Energy contributions and total
+    type(TEnergies), intent(inout) :: energy
+
+    !> energes of all solved states
+    real(dp), intent(inout), allocatable :: energies(:)
+
+    !> Working array of the size of the dense matrices.
+    real(dp), intent(out) :: work(:,:)
+
+    !> density matrix in dense form
+   !real(dp), intent(inout), allocatable :: rhoSqrReal(:,:,:)
+    real(dp), intent(inout) :: deltaRhoOutSqr(:,:,:)
+
+    !> excited state energy derivative with respect to atomic coordinates
+    real(dp), intent(inout), allocatable :: excitedDerivs(:,:)
+
+    !> Mulliken charges in excited state
+    real(dp), intent(out) :: dQAtomEx(:)
+
+    !> natural orbital occupation numbers
+    real(dp), intent(inout), allocatable :: occNatural(:)
+
+    !> Data from rangeseparated calculations
+    type(TRangeSepFunc), intent(inout), allocatable :: rangeSep
+
+
+    real(dp), allocatable :: dQAtom(:)
+    real(dp), allocatable :: naturalOrbs(:,:,:)
+    integer, pointer :: pSpecies0(:)
+    integer :: iSpin, nSpin, nAtom, fdAutotest
+    logical :: tSpin
+    ! Onsite corrections -- remain dummy as they are not calculated
+    logical, parameter :: tOnsite = .false.
+    real(dp), allocatable :: ons_en(:,:), ons_dip(:,:)
+    integer :: i, norb
+
+    nAtom = size(qOutput, dim=2)
+    nSpin = size(eigen, dim=2)
+    tSpin = (nSpin == 2)
+    pSpecies0 => species(1:nAtom)
+
+    norb = size(eigvecsReal, dim=1)
+
+    energy%Eexcited = 0.0_dp
+    allocate(dQAtom(nAtom))
+    dQAtom(:) = sum(qOutput(:,:,1) - q0(:,:,1), dim=1)
+    call unpackHS(work, over, neighbourList%iNeighbour, nNeighbourSK, denseDesc%iAtomStart,&
+        & iSparseStart, img2CentCell)
+    call blockSymmetrizeHS(work, denseDesc%iAtomStart)
+    !call unpackHS(SSqrReal, over, neighbourList%iNeighbour, nNeighbour,&
+    !     & denseDesc%iAtomStart, iPair, img2CentCell)
+    !call blockSymmetrizeHS(SSqrReal, iAtomStart)
+    if (tForces) then
+      do iSpin = 1, nSpin
+        call blockSymmetrizeHS(deltaRhoOutSqr(:,:,iSpin), denseDesc%iAtomStart)
+      end do
+    end if
+
+    if (tWriteAutotest) then
+      open(newUnit=fdAutotest, file=autotestTag, position="append")
+    end if
+
+    if (tLinRespZVect) then
+      if (tPrintExcEigVecs) then
+        allocate(naturalOrbs(orb%nOrb, orb%nOrb, 1))
+      end if
+      ! WITH FORCES
+      excitedDerivs = 0.0_dp
+    #:if WITH_ARPACK
+      call linRespCalcExcitationsRS(tSpin, tOnsite, linearResponse, denseDesc%iAtomStart,&
+          & eigvecsReal, eigen, sccCalc, work, filling, coord0, dQAtom, pSpecies0,&
+          & linearResponse%HubbardU, neighbourList%iNeighbour, img2CentCell, orb, rangeSep,&
+          & tWriteAutotest, fdAutotest, taggedWriter, energy%Eexcited, skHamCont, skOverCont,&
+          & nonSccDeriv, deltaRhoOutSqr(:,:,1), excitedDerivs, dQAtomEx)
+    #:else
+      call error("Should not be here - compiled without ARPACK")
+    #:endif
+      if (tPrintExcEigvecs) then
+        call writeRealEigvecs(env, runId, neighbourList, nNeighbourSK, denseDesc, iSparseStart,&
+            & img2CentCell, pSpecies0, speciesName, orb, over, parallelKS, tPrintExcEigvecsTxt,&
+            & naturalOrbs, work, fileName="excitedOrbs")
+      end if
+    else
+      ! NO FORCES
+    #:if WITH_ARPACK
+      call linRespCalcExcitationsRS(tSpin, tOnsite, linearResponse, denseDesc%iAtomStart,&
+          & eigvecsReal, eigen, sccCalc, work, filling, coord0, dQAtom, pSpecies0,&
+          & linearResponse%HubbardU, neighbourList%iNeighbour, img2CentCell, orb, rangeSep,&
+          & tWriteAutotest, fdAutotest, taggedWriter, energy%Eexcited)
+    #:else
+      call error("Should not be here - compiled without ARPACK")
+    #:endif
+    end if
+
+    energy%Etotal = energy%Etotal + energy%Eexcited
+    energy%EMermin = energy%EMermin + energy%Eexcited
+    energy%EGibbs = energy%EGibbs + energy%Eexcited
+    if (tWriteAutotest) then
+      close(fdAutotest)
+    end if
+
+  end subroutine calculateLinRespExcitations_RS
 
   !> Get the XLBOMD charges for the current geometry.
   subroutine getXlbomdCharges(xlbomdIntegrator, qOutRed, pChrgMixer, orb, nIneqOrb, iEqOrbitals,&
@@ -4923,7 +5097,7 @@ contains
     class(TDispersionIface), intent(inout), allocatable :: dispersion
 
     !> Data from rangeseparated calculations
-    type(TRangeSepFunc), intent(inout), allocatable ::rangeSep
+    type(TRangeSepFunc), intent(inout), allocatable :: rangeSep
 
     !> dense overlap matrix, required for rangeSep
     real(dp), intent(inout), allocatable :: SSqrReal(:,:)
@@ -5176,7 +5350,8 @@ contains
   subroutine getStress(env, sccCalc, thirdOrd, tExtField, nonSccDeriv, rhoPrim, ERhoPrim, qOutput,&
       & q0, skHamCont, skOverCont, pRepCont, neighbourList, nNeighbourSk, nNeighbourRep, species,&
       & img2CentCell, iSparseStart, orb, potential, coord, latVec, invLatVec, cellVol, coord0,&
-      & totalStress, totalLatDeriv, intPressure, iRhoPrim, solvation, dispersion, halogenXCorrection)
+      & totalStress, totalLatDeriv, intPressure, iRhoPrim, solvation, dispersion,&
+      & halogenXCorrection)
 
     !> Environment settings
     type(TEnvironment), intent(in) :: env
@@ -5916,7 +6091,7 @@ contains
 
   subroutine sendEnergyAndForces(env, socket, energy, TS, derivs, totalStress, cellVol)
     type(TEnvironment), intent(in) :: env
-    ! Socket may be unallocated (as on slave processes)
+    ! Socket may be unallocated (as on follower processes)
     type(ipiSocketComm), allocatable, intent(inout) :: socket
     type(TEnergies), intent(in) :: energy
     real(dp), intent(in) :: TS(:)
@@ -5924,7 +6099,7 @@ contains
     real(dp), intent(in) :: totalStress(:,:)
     real(dp), intent(in) :: cellVol
 
-    if (env%tGlobalMaster) then
+    if (env%tGlobalLead) then
       ! stress was computed above in the force evaluation block or is 0 if aperiodic
       call socket%send(energy%ETotal - sum(TS), -derivs, totalStress * cellVol)
     end if