+++ prog/dftb+/lib_dftbplus/parser.F90
@@ -59,6 +59,7 @@ module dftbp_parser
   use dftbp_tempprofile, only : identifyTempProfile
   use dftbp_reks
   use dftbp_plumed, only : withPlumed
+  use dftbp_arpack, only : withArpack
   use poisson_init
 #:if WITH_TRANSPORT
   use libnegf_vars
@@ -1210,7 +1211,7 @@ contains
     integer :: nElem
     real(dp) :: rSKCutOff
 
-    !>For rangeseparation
+    !> For range separation
     logical :: tRSep
     real(dp) :: screeningThreshold
     type(TRangeSepSKTag) :: rangeSepSK
@@ -4341,22 +4342,17 @@ contains
 
   #:if WITH_ARPACK
     type(string) :: modifier
-  #:endif
 
     ! Linear response stuff
     call getChild(node, "Casida", child, requested=.false.)
 
-  #:if not WITH_ARPACK
-
-    if (associated(child)) then
+    if (associated(child) .and. .not. withArpack) then
       call detailedError(child, 'This DFTB+ binary has been compiled without support for linear&
-          & response calculations (requires the ARPACK/ngARPACK library).')
+          & response calculations (requires the ARPACK/ngARPACK libraries).')
     end if
 
     ctrl%lrespini%tInit = .false.
     ctrl%lrespini%tPrintEigVecs = .false.
-
-  #:else
 
     if (associated(child)) then
 
@@ -4377,6 +4373,10 @@ contains
           call detailedError(child2, "Invalid symmetry value '"  // char(buffer) // &
               & "' (must be 'Singlet', 'Triplet' or 'Both').")
         end select
+        if (allocated(ctrl%rangeSepInp) .and. ctrl%lrespini%sym /= 'S') then
+           call detailedError(child2, "Invalid symmetry value '"  // char(buffer) // &
+              & "' (only 'Singlet' for LC-DFTB).")
+        endif
       end if
 
       call getChildValue(child, "NrOfExcitations", ctrl%lrespini%nexc)
@@ -4425,6 +4425,9 @@ contains
       call getChildValue(child, "WriteSPTransitions", ctrl%lrespini%tSPTrans, default=.false.)
       call getChildValue(child, "WriteTransitions", ctrl%lrespini%tTrans, default=.false.)
       call getChildValue(child, "WriteTransitionDipole", ctrl%lrespini%tTradip, default=.false.)
+      if (allocated(ctrl%rangeSepInp)) then
+        call getChildValue(child, "WriteTransitionCharges", ctrl%lrespini%tTransQ, default=.false.)
+      end if
       call getChildValue(child, "WriteStatusArnoldi", ctrl%lrespini%tArnoldi, default=.false.)
       call getChildValue(child, "TestArnoldi", ctrl%lrespini%tDiagnoseArnoldi, default=.false.)
 
@@ -4469,7 +4472,7 @@ contains
       call getChildValue(child, "TammDancoff", ctrl%pprpa%tTDA, default=.false.)
 
       call getChild(child, "NrOfVirtualStates", child2, requested=.false.)
-      if (.not. associated(child2)) then      
+      if (.not. associated(child2)) then
         ctrl%pprpa%nvirtual = 0
         ctrl%pprpa%tConstVir = .false.
         call setChildValue(child, "NrOfVirtualStates", 0)